requirements.txt
streamlit
pymupdf
Pillow
opencv-python
pandas